Almacenar información adecuada de los Usuarios
